Tabs Configuration
Created a tabbed layout within the tabs folder. Most work was done in the _layout file which mainly just formatted the bottom tabs bar.

It added a Tab Icon react functional component which then allows up to create 4 same tab compoenents withing the <Tabs> component

Each component had the tabs.screen component attatched which told it to create a screen based on that componenet that exists in the (tabs) folder and the TabIcon Component inside it provided the icon name and alignment

At least thats how I understand that it works......