Personalizes Sign in and Sign Up
The sign in and sign up function now use the text boxes in the sign in and sign up page. Before it was using the built in function that was sending pre build values but now we have linked the form fields submission to the database submission.

the first submit fucntion was created in the sign up file and that same one was copied and implemented in the sign in file

for now I still have to go into appwrite -> auth -> sessions and delete the session but will figure out how to resolve that later.